bhyve/amd64: Add support for CPUID configuration
Generate CPUID entries from any given CPUID config options and populate
the CPUID configuration for all VCPUs accordingly.

CPUID options conform to the following form:

-o [vcpu.<vcpuid>.]cpuid.<leaf>[,<index>]=<eax>,<ebx>,<ecx>,<edx>

<leaf> specifies the CPUID leaf (aka "CPUID function") to create an
entry for, with an optional index specified by <index>. Leaf and index
correspond the EAX and ECX input register values, respectively, when a
CPUID instruction is emulated for the guest. <eax>, <ebx>, <ecx>, and
<edx> specify the corresponding register values to be returned for the
specified leaf and index.

When a CPUID option is rooted under 'cpuid.' in the config tree, it
defines an entry for the global CPUID configuration, which is applied
to all VCPUs. Any CPUID option rooted under 'vcpu.<vcpuid>.cpuid.'
defines a CPUID configuration entry specifically for the given VCPU.

Since vmm expects to handle only VCPU-specific CPUID configurations,
the global CPUID configuration is merged into the VCPU-specific CPUID
configuration, but VCPU-specific CPUID configuration options always
override any global CPUID options for the same leaf/index pair.

CPUID options can be specified on the command line in any order.
Duplicates aren't allowed, although detection of duplicates is limited
by the mechanism of which '-o' options are parsed in bhyve.

In particular, consider the following cases:
1: bhyve [...] -o cpuid.0x123=0,0,0,0 -o cpuid.0x123=1,2,3,4 [...]
2: bhyve [...] -o cpuid.0x00000123=0,0,0,0 -o cpuid.0x123=1,2,3,4 [...]

The first case isn't considered duplicate entry as the bhyve option
parsing code only keeps the last value given for a particular option
name string, which in both cases is 'cpuid.0x123'.

In the 2nd case, the option name string is different, thus they are two
separate options as far as the bhyve option parsing code is concerned
and both will be entered into the config nvlist. But the cpuid option
parser decodes both options into two different cpuid entries for the
same cpuid leaf, which is an error and reported as such.

Note that if a cpuid configuration is to be used, all CPUID leaves and
indices must be specified, including all VCPU-specific entries where
they differ from one VCPU to another (such as for the APIC ID, CPU
topology information etc.). Eventually, we'd like to have a way to
override only individual CPUID leaves or even just masking/setting
individual bits in a register value, falling back to the legacy CPUID
emulation for leaves and indices not explicitly configured.